customizable PreviewCard (#435)
* customizable PreviewCard

* fix className props type
