+++ src/components/PreviewCard/PreviewCard.styles.ts
@@ -1,14 +1,14 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
-import { getTheme } from '../../util/colors';
-import { mergeStyleSets } from '@uifabric/styling';
+import { mergeStyleSets, ITheme } from '@uifabric/styling';
 import { memoizeFunction } from '@uifabric/utilities';
 
-export interface PreviewCardStyleProps {
+export interface PreviewCardClassNameProps {
   allowOnClick: boolean;
+  theme: ITheme;
 }
 
-const getMemoizedClassNames = memoizeFunction((styleProps, theme) => {
-  const { allowOnClick } = styleProps;
+export const getClassNames = memoizeFunction((styleProps: PreviewCardClassNameProps) => {
+  const { allowOnClick, theme } = styleProps;
 
   return mergeStyleSets({
     root: {
@@ -35,8 +35,3 @@ const getMemoizedClassNames = memoizeFunction((styleProps, theme) => {
     },
   });
 });
-
-export const getClassNames = (styleProps: PreviewCardStyleProps) => {
-  const theme = getTheme();
-  return getMemoizedClassNames(styleProps, theme);
-};
