+++ src/components/PreviewCard/PreviewCard.test.tsx
@@ -1,10 +1,12 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
 import * as React from 'react';
-import { shallow, ShallowWrapper } from 'enzyme';
+import { shallow, mount, ShallowWrapper, ReactWrapper } from 'enzyme';
 import Clickable from '../Clickable';
 import EditableText from '../EditableText';
-import PreviewCard, { PreviewCardProps } from './index';
+import CustomizablePreviewCard, { PreviewCard } from './PreviewCard';
+import { PreviewCardProps } from './PreviewCard.types';
 import { renderIntoDocument, findRenderedComponentWithType } from 'react-dom/test-utils';
+import Customizer, { defaultTheme } from '../Customizer';
 
 describe('<PreviewCard />', () => {
   let component: ShallowWrapper<PreviewCardProps>;
@@ -176,6 +178,23 @@ describe('<PreviewCard />', () => {
       it('triggers the onRemoveClick callback prop', () => {
         expect(onClick).toHaveBeenCalledTimes(1);
       });
+    });
+  });
+
+  describe('with customizer', () => {
+    let mountedComponent: ReactWrapper;
+    const theme = defaultTheme;
+
+    beforeEach(() => {
+      mountedComponent = mount(
+        <Customizer settings={{ theme }}>
+          <CustomizablePreviewCard name="Filename.gif" />
+        </Customizer>,
+      );
+    });
+
+    it('receives custom theme', () => {
+      expect(mountedComponent.find('CustomizablePreviewCard').prop('theme')).toBe(theme);
     });
   });
 });
