+++ src/components/PreviewCard/PreviewCard.tsx
@@ -1,7 +1,7 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
 import '../../yamui';
 import * as React from 'react';
-import { BaseComponentProps } from '../../util/BaseComponent/props';
+import { CustomizableComponentProps, defaultTheme, customizable } from '../Customizer';
 import { join } from '../../util/classNames';
 import Block, { GutterSize, TextColor, TextSize } from '../Block';
 import Box from '../Box';
@@ -14,98 +14,13 @@ import MediaObject, { MediaObjectProps, MediaObjectSize } from '../MediaObject';
 import ProgressIndicator from '../ProgressIndicator';
 import Spinner, { SpinnerColor, SpinnerSize } from '../Spinner';
 import { getClassNames } from './PreviewCard.styles';
-
-export interface PreviewCardProps extends BaseComponentProps {
-  /**
-   * The name of the file or entity being represented. This will be displayed as the title.
-   */
-  name: string;
-
-  /**
-   * Description of the file.
-   */
-  description?: string;
-
-  /**
-   * Whether the description can be edited or not.
-   */
-  isDescriptionEditable?: boolean;
-
-  /**
-   * Will display on the left of the PreviewCard. If not provided the image will be replaced
-   * by an icon depending on the file type.
-   */
-  imageUrl?: string;
-
-  /**
-   * The alt text to be displayed for the image.
-   */
-  imageDescription?: string;
-
-  /**
-   * If true will display a Spinner, or a ProgressIndicator if progress is also provided.
-   * Note that the loading Spinner will be displayed in place of the description.
-   */
-  isLoading?: boolean;
-
-  /**
-   * A string to describe the loading state. If you've provided a progress value then
-   * this string should also include that percentage.
-   */
-  loadingText?: string;
-
-  /**
-   * Percentage between 0 and 1. If provided, will replace the loading Spinner with
-   * a ProgressIndicator to show the percentage.
-   */
-  progress?: number;
-
-  /**
-   * The text to display when the description is editable and currently empty.
-   */
-  emptyEditableDescriptionText?: string;
-
-  /**
-   * An optional max length for the description field when editing.
-   */
-  descriptionMaxLength?: number;
-
-  /**
-   * Triggered when the PreviewCard is clicked. This will not be triggered for clicks on the Remove icon
-   * or to edit the description.
-   */
-  onClick?: (() => void);
-
-  /**
-   * AriaLabel value describing the onClick action.
-   */
-  clickAriaLabel?: string;
-
-  /**
-   * Triggered when the Remove icon is clicked.
-   */
-  onRemoveClick?: (() => void);
-
-  /**
-   * Alt text for the Remove button
-   */
-  removeAriaLabel?: string;
-
-  /**
-   * Returns the new description string when updated.
-   */
-  onDescriptionChange?: ((description: string) => void);
-}
-
-export interface PreviewCardState {
-  isEditing: boolean;
-}
+import { PreviewCardProps, PreviewCardState } from './PreviewCard.types';
 
 /**
  * PreviewCard is a compact representation of an uploaded file or other attachment. It supports an
  * editable description, and a loading experience (Spinner or percentage ProgressIndicator).
  */
-export default class PreviewCard extends React.Component<PreviewCardProps, PreviewCardState> {
+export class PreviewCard extends React.Component<PreviewCardProps & CustomizableComponentProps, PreviewCardState> {
   constructor(props: PreviewCardProps) {
     super(props);
     this.state = {
@@ -123,8 +38,9 @@ export default class PreviewCard extends React.Component<PreviewCardProps, Previ
       imageUrl,
       isLoading,
       imageDescription = '',
+      theme = defaultTheme,
     } = this.props;
-    const classNames = getClassNames({ allowOnClick: this.allowOnClick() });
+    const classNames = getClassNames({ allowOnClick: this.allowOnClick(), theme });
 
     const mediaObjectProps: MediaObjectProps = {
       size: MediaObjectSize.MEDIUM,
@@ -239,3 +155,6 @@ export default class PreviewCard extends React.Component<PreviewCardProps, Previ
     return typeof this.props.progress === 'number';
   }
 }
+
+@customizable('PreviewCard', ['theme'])
+export default class CustomizablePreviewCard extends PreviewCard {}
